requirements.txt
africastalking==1.2.9
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.5
django-celery-beat==2.7.0
django-timezone-field==7.1
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
kombu==5.4.2
multidict==6.1.0
mysqlclient==2.2.7
prompt_toolkit==3.0.50
propcache==0.2.1
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
schema==0.7.7
six==1.17.0
sqlparse==0.5.3
twilio==9.4.4
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.18.3
